Review comments and moving the constructions of schedulers in their own factory where BlazeRunner could construct them.
