+++ tt-media-server/cpp_server/CMakeLists.txt
@@ -466,6 +466,7 @@ if(ENABLE_BLAZE)
     list(APPEND LLM_RUNNER_LIB_SOURCES
         src/runtime/runners/blaze_runner/blaze_decode_runner.cpp
         src/runtime/runners/blaze_runner/blaze_prefill_runner.cpp
+        src/runtime/runners/blaze_runner/blaze_scheduler_factory.cpp
     )
     message(STATUS "llm_runner_lib: including tt-llm-engine-dependent sources (ENABLE_BLAZE=ON)")
 endif()
