+++ tt-media-server/cpp_server/include/config/defaults.hpp
@@ -148,7 +148,7 @@ constexpr int64_t DYNAMO_ETCD_LEASE_TTL_SECS = 10;
 
 // MockSchedulers (MOCK_USE_SCHEDULER=1 on mock_pipeline backend).
 constexpr bool MOCK_USE_SCHEDULER = false;
-constexpr unsigned MOCK_PREFILL_LATENCY_MS = 1353;
+constexpr unsigned MOCK_PREFILL_CHUNK_LATENCY_MS = 1353;
 constexpr unsigned MOCK_DECODE_TOKEN_LATENCY_US =
     64 * 44;  // 64 * 44us = 2816us
 constexpr unsigned MOCK_DECODE_TOKEN_ID = 12345;
