+++ tt-media-server/cpp_server/include/runtime/runners/blaze_runner/blaze_scheduler_factory.hpp
@@ -1,0 +1,19 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
+
+#pragma once
+
+#include <memory>
+
+#include "config/runner_config.hpp"
+#include "runtime/runners/blaze_runner/scheduler_interface.hpp"
+
+namespace tt::runners::blaze {
+
+std::unique_ptr<IDecodeScheduler> makeDecodeScheduler(
+    const tt::config::LLMConfig& config);
+
+std::unique_ptr<IPrefillScheduler> makePrefillScheduler(
+    const tt::config::LLMConfig& config);
+
+}  // namespace tt::runners::blaze
