+++ tt-media-server/cpp_server/src/runtime/runners/blaze_runner/blaze_decode_runner.cpp
@@ -15,96 +15,29 @@
 #include "ipc/helpers/token_push.hpp"
 #include "runtime/runners/blaze_runner/blaze_slot_manager.hpp"
 #include "runtime/runners/blaze_runner/blaze_utils.hpp"
-#include "runtime/runners/blaze_runner/mock_scheduler.hpp"
 #include "runtime/worker/single_process_worker_metrics.hpp"
 #include "services/memory_services/memory_manager.hpp"
-#include "tt_llm_engine/scheduler/decode/decode_scheduler.hpp"
 #include "tt_llm_engine/scheduler/scheduler_types.hpp"
 #include "utils/logger.hpp"
-#include "utils/tokenizers/tokenizer.hpp"
 
 namespace tt::runners::blaze {
-namespace {
-
-class RealDecodeScheduler final : public IDecodeScheduler {
- public:
-  explicit RealDecodeScheduler(std::unique_ptr<ds::DecodeScheduler> scheduler)
-      : impl(std::move(scheduler)) {}
-
-  void start() override { impl->start(); }
-  void stop() override { impl->stop(); }
-  bool push_request(const ds::ISRequest& request) override {
-    return impl->push_request(request);
-  }
-  bool try_pop_response(ds::SchedulerResponse& response) override {
-    return impl->try_pop_response(response);
-  }
-  bool try_pop_output(ds::OutputMessage& output) override {
-    return impl->try_pop_output(output);
-  }
-  uint32_t get_spec_accepts(uint32_t slotId) const override {
-    return impl->get_spec_accepts(slotId);
-  }
-  uint32_t get_spec_rejects(uint32_t slotId) const override {
-    return impl->get_spec_rejects(slotId);
-  }
-
- private:
-  std::unique_ptr<ds::DecodeScheduler> impl;
-};
-
-}  // namespace
 
 namespace sched = tt_llm_engine::scheduler;
 BlazeDecodeRunner::BlazeDecodeRunner(
-    const config::LLMConfig& config, ipc::IResultQueue* resultQueue,
-    tt::ipc::ITaskQueue* taskQueue, tt::ipc::ICancelQueue* stopQueue,
+    const config::LLMConfig& config,
+    std::unique_ptr<IDecodeScheduler> decodeScheduler,
+    ipc::IResultQueue* resultQueue, tt::ipc::ITaskQueue* taskQueue,
+    tt::ipc::ICancelQueue* stopQueue,
     std::unique_ptr<tt::services::MemoryManager> injectedMemoryManager)
     : config(config),
       resultQueue(resultQueue),
       taskQueue(taskQueue),
       stopQueue(stopQueue),
+      decodeScheduler(std::move(decodeScheduler)),
       slotManager(tt::config::pmMaxUsers()),
       outputHangTimeout(tt::config::outputHangTimeoutMs()) {
-  const auto maxUsers = static_cast<uint32_t>(tt::config::pmMaxUsers());
-  if (config.runner_type == tt::config::ModelRunnerType::MOCK_PIPELINE &&
-      tt::config::useMockScheduler()) {
-    TT_LOG_INFO(
-        "BlazeDecodeRunner: using MockDecodeScheduler (single-threaded)");
-    decodeScheduler = std::make_unique<MockDecodeScheduler>(maxUsers);
-  } else {
-    TT_LOG_INFO(
-        "BlazeDecodeRunner: Constructing DecodeScheduler with SocketConfig...");
-    auto pipelineConfig = utils::makeDecodePipelineConfig(config);
-    auto migrationClientInterface =
-        utils::makeDecodeMigrationClientInterface(config);
-    auto thinkTokenIds = tt::utils::tokenizers::thinkTokenIds();
-    auto eosTokenId = tt::utils::tokenizers::staticInfo().eosTokenId;
-    ds::SchedulerParams managerParams{};
-    managerParams.num_layers = tt::config::modelNumLayers();
-    managerParams.eos_token = static_cast<uint32_t>(eosTokenId);
-    managerParams.think_open_token_id =
-        static_cast<uint32_t>(thinkTokenIds.first);
-    managerParams.think_close_token_id =
-        static_cast<uint32_t>(thinkTokenIds.second);
-    managerParams.max_users = maxUsers;
-    managerParams.self_endpoint_id = tt::config::migrationDecodeEndpointId();
-    if (tt::config::enableMigration()) {
-      migrationClientInterface->connect_to(
-          tt::config::migrationPrefillEndpointId(), "CONNECTOR", "ds_pd");
-    }
-    if (tt::config::specDecodeMode() == "mtp") {
-      managerParams.spec_decode_mode = ds::SpecDecodeMode::MTP;
-      managerParams.max_spec_tokens =
-          static_cast<uint32_t>(tt::config::mtpLevel());
-    }
-    decodeScheduler = std::make_unique<RealDecodeScheduler>(
-        std::make_unique<ds::DecodeScheduler>(
-            pipelineConfig, managerParams,
-            std::move(migrationClientInterface)));
-    TT_LOG_INFO("BlazeDecodeRunner: DecodeScheduler constructed");
-  }
-  decodeScheduler->start();
+  assert(this->decodeScheduler != nullptr);
+  this->decodeScheduler->start();
   TT_LOG_INFO(
       "BlazeDecodeRunner: PipelineManager started, creating MemoryManager...");
   memoryManager = injectedMemoryManager
