+++ tt-media-server/cpp_server/src/runtime/runners/blaze_runner/blaze_prefill_runner.cpp
@@ -15,81 +15,28 @@
 #include "ipc/helpers/token_push.hpp"
 #include "runtime/runners/blaze_runner/blaze_slot_manager.hpp"
 #include "runtime/runners/blaze_runner/blaze_utils.hpp"
-#include "runtime/runners/blaze_runner/mock_scheduler.hpp"
 #include "runtime/worker/single_process_worker_metrics.hpp"
 #include "services/memory_services/memory_manager.hpp"
-#include "tt_llm_engine/scheduler/prefill/prefill_scheduler.hpp"
 #include "utils/logger.hpp"
 
 namespace tt::runners::blaze {
-namespace {
-
-namespace ps = tt_llm_engine::scheduler::prefill;
-
-class RealPrefillScheduler final : public IPrefillScheduler {
- public:
-  explicit RealPrefillScheduler(std::unique_ptr<ps::PrefillScheduler> scheduler)
-      : impl(std::move(scheduler)) {}
-
-  void start() override { impl->start(); }
-  void stop() override { impl->stop(); }
-  bool push_request(const ps::ISRequest& request) override {
-    return impl->push_request(request);
-  }
-  bool try_pop_response(ps::SchedulerResponse& response) override {
-    return impl->try_pop_response(response);
-  }
-  bool try_pop_output(ps::OutputMessage& output) override {
-    return impl->try_pop_output(output);
-  }
-
- private:
-  std::unique_ptr<ps::PrefillScheduler> impl;
-};
-
-}  // namespace
 
 BlazePrefillRunner::BlazePrefillRunner(
-    const config::LLMConfig& config, ipc::IResultQueue* resultQueue,
-    tt::ipc::ITaskQueue* taskQueue, tt::ipc::ICancelQueue* stopQueue,
+    const config::LLMConfig& config,
+    std::unique_ptr<IPrefillScheduler> prefillScheduler,
+    ipc::IResultQueue* resultQueue, tt::ipc::ITaskQueue* taskQueue,
+    tt::ipc::ICancelQueue* stopQueue,
     std::unique_ptr<tt::services::MemoryManager> injectedMemoryManager)
     : config(config),
       resultQueue(resultQueue),
       taskQueue(taskQueue),
       stopQueue(stopQueue),
+      prefillScheduler(std::move(prefillScheduler)),
       slotManager(tt::config::pmMaxUsers()),
       lastOutputTime(std::chrono::steady_clock::now()),
       outputHangTimeout(tt::config::outputHangTimeoutMs()) {
-  const auto maxUsers = static_cast<uint32_t>(tt::config::pmMaxUsers());
-  if (config.runner_type == tt::config::ModelRunnerType::MOCK_PIPELINE &&
-      tt::config::useMockScheduler()) {
-    TT_LOG_INFO(
-        "BlazePrefillRunner: using MockPrefillScheduler (single-threaded)");
-    prefillScheduler = std::make_unique<MockPrefillScheduler>(maxUsers);
-  } else {
-    TT_LOG_INFO(
-        "BlazePrefillRunner: Constructing PrefillScheduler with "
-        "SocketConfig...");
-    auto pipelineConfig = utils::makePrefillPipelineConfig(config);
-    ps::SchedulerParams managerParams{};
-    managerParams.dest_endpoint_id = tt::config::migrationDecodeEndpointId();
-    managerParams.self_endpoint_id = tt::config::migrationPrefillEndpointId();
-    managerParams.layers_per_chunk = tt::config::modelNumLayers();
-    managerParams.chunk_size = tt::config::prefillChunkSize();
-    managerParams.max_users = maxUsers;
-    auto ackChannelConfig = utils::makePrefillAckChannelConfig(config);
-    auto migrationClientInterface = utils::makeMigrationClientInterface(config);
-    if (tt::config::enableMigration()) {
-      migrationClientInterface->connect_to(
-          tt::config::migrationDecodeEndpointId(), "PUBLISHER", "ds_pd");
-    }
-    prefillScheduler = std::make_unique<RealPrefillScheduler>(
-        std::make_unique<ps::PrefillScheduler>(
-            pipelineConfig, ackChannelConfig, managerParams,
-            std::move(migrationClientInterface)));
-    TT_LOG_INFO("BlazePrefillRunner: PrefillScheduler constructed");
-  }
-  prefillScheduler->start();
+  assert(this->prefillScheduler != nullptr);
+  this->prefillScheduler->start();
   TT_LOG_INFO(
       "BlazePrefillRunner: PipelineManager started, creating MemoryManager...");
   memoryManager = injectedMemoryManager
